♻️ Refactor image handler
